+++ src/lakebench/engines/duckdb.py
@@ -33,16 +33,6 @@ class DuckDB(BaseEngine):
 
         self.version: str = f"{version('duckdb')} (deltalake=={version('deltalake')})"
         self.cost_per_vcore_hour = cost_per_vcore_hour or getattr(self, '_FABRIC_USD_COST_PER_VCORE_HOUR', None)
-
-    def create_schema_if_not_exists(self, drop_before_create: bool = True):
-        if drop_before_create:
-            try:
-                self.notebookutils.fs.rm(self.delta_abfss_schema_path, recurse=True)
-            except FileNotFoundError:
-                pass
-            except Exception as e:
-                raise e
-        # no need to create schema for Python engines
     
     def _create_empty_table(self, table_name: str, ddl: str):
         if not ddl.strip().startswith("CREATE OR REPLACE TABLE"):
